Defined class for 1qb RB + cleaned the scripts
