+++ bakery/baking_usage_examples/Ramsey_fringes/RamseyGauss_configuration.py
